Rename snake_case for position_dropdown
